Moves ELF handling to kernel
