+++ src/Cargo.toml
@@ -2,7 +2,6 @@
 resolver = "2"
 members = [
     "core",
-    "elf",
     "error",
     "fs",
     "ftp",
