+++ src/kernel/Cargo.toml
@@ -13,7 +13,6 @@ bytemuck = "1.14.0"
 byteorder = "1.4"
 clap = { version = "4.1", features = ["derive"] }
 discord-rich-presence = "0.2"
-elf = { path = "../elf" }
 gmtx = { path = "../gmtx" }
 iced-x86 = { version = "1.18", features = ["code_asm"] }
 libc = "0.2"
